Pin the tree-sitter CLI so a stale one cannot downgrade src/
`src/parser.c`, `src/grammar.json`, `src/node-types.json` and `src/tree_sitter/`
are generated and committed, so the CLI version is part of the build. Debian
ships `tree-sitter-cli` 0.20.8 at /usr/bin/tree-sitter, and a bare
`tree-sitter generate` picked it up in any shell where nvm had not loaded,
rewriting all four files in the pre-0.25 format — a ~150k-line diff that is
easy to commit by accident.

Nothing resolves a bare `tree-sitter` off PATH any more. package.json pins
tree-sitter-cli, run-tests.sh installs it on first run, refuses to proceed if
the resolved binary is not the pinned major.minor, and reports (rather than
hides) a regeneration that moves src/. test/highlights/check.py resolves the
same binary instead of shelling out to whatever is first on PATH, which is what
made the suite fail outside an nvm shell. package-lock.json is committed so the
pin is exact, and .npmrc allows the tree-sitter-cli install script that npm 11
now gates by default.

The generated output is unchanged: 0.26.12 reproduces the committed
src/parser.c byte for byte, and the only src/ delta is the vendored
tree_sitter/array.h that ships with the newer CLI. Deliberately still ABI 14 —
the absence of tree-sitter.json is what pins that, and prettier-move loads this
grammar through web-tree-sitter 0.20.x, which cannot load ABI 15. README now
says so where someone would look before "fixing" the warning.

Also refreshed the stale dependency floors (node-addon-api 7 -> 8,
peerDependency tree-sitter 0.21 -> 0.25, prebuildify, Cargo dev-dep
tree-sitter 0.25 -> 0.26), fixed `npm test` pointing at a nonexistent `test`
binary and the `tree_sitter` typo in peerDependenciesMeta, and adjusted the
Rust test for tree-sitter 0.26's `Node::child(u32)`.

Verified: corpus 111 pass / 1 skip / 0 fail (was 100 pass / 12 fail purely
because 0.20.8 could not parse the corpus `:skip` attribute), highlight
assertions OK, cargo tests pass, parse sweep clean, tree sha256 identical
across 16,266 real .move files, and the whole suite green in a shell where
/usr/bin/tree-sitter 0.20.8 is the only tree-sitter on PATH.

+++ bindings/rust/lib.rs
@@ -69,7 +69,7 @@ mod tests {
                 if node.kind() == "doc_comment" {
                     widths.push(node.end_byte() - node.start_byte());
                 }
-                for i in 0..node.child_count() {
+                for i in 0..node.child_count() as u32 {
                     stack.push(node.child(i).unwrap());
                 }
             }
